requirements.txt
annotated-types==0.7.0
ansible==10.7.0
ansible-core==2.17.13
ansible-runner==2.4.1
anyio==3.7.1
APScheduler==3.11.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cryptography==45.0.6
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.10
Jinja2==3.1.6
lockfile==0.12.2
MarkupSafe==3.0.2
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
pandas==2.1.4
paramiko==3.3.1
passlib==1.7.4
pexpect==4.9.0
ptyprocess==0.7.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyMySQL==1.1.1
PyNaCl==1.5.0
python-daemon==3.1.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.1
requests==2.32.5
resolvelib==1.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.42
starlette==0.27.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
XlsxWriter==3.1.9
